django_src/requirements.txt
certifi==2018.8.24
Django==2.0
django-appconf==1.0.2
django-bootstrap-form==3.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
